backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
chromadb==0.5.20
PyPDF2==3.0.1
pycryptodome==3.21.0
google-generativeai==0.8.4
python-multipart==0.0.20
indic-transliteration==2.3.57
rapidfuzz==3.12.1
pytest==8.3.4
httpx==0.28.1
